+++ upcloud/testdata/upcloud_loadbalancer/loadbalancer_e1.tf
@@ -1,13 +1,18 @@
 // Actions validation: Empty actions array
 
-variable "lb_zone" {
+variable "basename" {
+  default = "tf-acc-test-lb-errors"
+  type    = string
+}
+
+variable "zone" {
   type    = string
   default = "fi-hel2"
 }
 
 resource "upcloud_network" "lb_network" {
-  name = "lb-test-net"
-  zone = var.lb_zone
+  name = "${var.basename}net"
+  zone = var.zone
   ip_network {
     address = "10.0.9.0/24"
     dhcp    = true
@@ -17,15 +22,27 @@ resource "upcloud_network" "lb_network" {
 
 resource "upcloud_loadbalancer" "lb" {
   configured_status = "started"
-  name              = "lb-test"
+  name              = "${var.basename}lb"
   plan              = "development"
-  zone              = var.lb_zone
-  network           = resource.upcloud_network.lb_network.id
+  zone              = var.zone
+
+  networks {
+    type   = "public"
+    family = "IPv4"
+    name   = "public"
+  }
+
+  networks {
+    type    = "private"
+    family  = "IPv4"
+    name    = "private"
+    network = resource.upcloud_network.lb_network.id
+  }
 }
 
 resource "upcloud_loadbalancer_backend" "lb_be_1" {
-  loadbalancer  = resource.upcloud_loadbalancer.lb.id
-  name          = "lb-be-test"
+  loadbalancer = resource.upcloud_loadbalancer.lb.id
+  name         = "lb-be-test"
 }
 
 resource "upcloud_loadbalancer_frontend" "lb_fe_1" {