+++ upcloud/testdata/upcloud_loadbalancer/loadbalancer_s3.tf
@@ -1,23 +1,25 @@
-variable "lb_zone" {
+variable "basename" {
+  default = "tf-acc-test-lb-"
+  type    = string
+}
+
+variable "zone" {
   type    = string
   default = "fi-hel2"
 }
 
-resource "upcloud_network" "lb_network" {
-  name = "lb-test-net"
-  zone = var.lb_zone
-  ip_network {
-    address = "10.0.7.0/24"
-    dhcp    = true
-    family  = "IPv4"
-  }
+locals {
+  addresses = ["10.0.7.0/24", "10.0.8.0/24"]
 }
 
-resource "upcloud_network" "lb_network_2" {
-  name = "lb-test-net-2"
-  zone = var.lb_zone
+resource "upcloud_network" "lb_network" {
+  count = length(local.addresses)
+
+  name = "${var.basename}net-${count.index}"
+  zone = var.zone
+
   ip_network {
-    address = "10.0.8.0/24"
+    address = local.addresses[count.index]
     dhcp    = true
     family  = "IPv4"
   }
@@ -25,23 +27,22 @@ resource "upcloud_network" "lb_network_2" {
 
 resource "upcloud_loadbalancer" "lb" {
   configured_status = "started"
-  name              = "lb-test"
+  name              = "${var.basename}lb"
   plan              = "development"
-  zone              = var.lb_zone
+  zone              = var.zone
   maintenance_dow   = "monday"
   maintenance_time  = "00:01:01Z"
-  # change: move from network ID to network objects forces replacement
-  networks {
-    name    = "lan-1"
-    type    = "private"
-    family  = "IPv4"
-    network = resource.upcloud_network.lb_network.id
-  }
-  networks {
-    name    = "lan-2"
-    type    = "private"
-    family  = "IPv4"
-    network = resource.upcloud_network.lb_network_2.id
+
+  # change: from public → private to private → private
+  dynamic "networks" {
+    for_each = upcloud_network.lb_network
+
+    content {
+      name    = "lan-${networks.key}"
+      type    = "private"
+      family  = "IPv4"
+      network = networks.value.id
+    }
   }
 }
 
@@ -51,13 +52,15 @@ resource "upcloud_loadbalancer_frontend" "lb_fe_1" {
   mode                 = "http"
   port                 = 8080
   default_backend_name = resource.upcloud_loadbalancer_backend.lb_be_1.name
+
   properties {
     timeout_client         = 20
     inbound_proxy_protocol = true
   }
+
   # change: add network listener
   networks {
-    name = resource.upcloud_loadbalancer.lb.networks[0].name
+    name = "lan-0"
   }
 }